# api/openapi.yaml
openapi: 3.0.3
info:
  title: Recipe Management API
  version: 1.0.0
  description: |
    レシピ管理アプリの API 契約。
    本ファイルを契約の単一ソースとし、Go(oapi-codegen)と
    TypeScript(フロント)双方の型を生成する。
servers:
  - url: /
    description: 同一オリジン(dev は Vite proxy で /api を転送、本番はリバースプロキシ)
...
paths:
...
  /api/share_group/invite_code/:
    post:
      tags: [share-group]
      summary: 招待コードを再発行する(所有者のみ。旧コードは失効)
      operationId: regenerateInviteCode
      responses:
        '200':
          description: 再発行成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShareGroupResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Authorization: Bearer <access_token>。
        refresh トークンの Cookie 化は別途対応予定。
  responses:
...
    Unauthorized:
      description: 未認証・トークン無効/期限切れ・クレデンシャル不正
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: 権限なし(他ユーザーのレシピ操作など)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: 対象が存在しない
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
...
    Error:
      type: object
      description: Echo の既定エラー形式
      required: [message]
      properties:
        message:
          type: string
          example: not found
...
    UserListItem:
      type: object
      required: [id, username, avatar_url]
      properties:
        id:
          type: string
          format: uuid
          x-go-name: ID
          x-go-type: string
        username:
          type: string
        avatar_url:
          type: string
          nullable: true
          description: プロフィール画像の URL。未設定なら null。
...
    ShareGroupResponse:
      type: object
      description: シェアグループ。members は所有者を含む全メンバー。is_owner は取得ユーザーが所有者か。
      required: [id, name, owner, members, invite_code, invite_code_expires_at, is_owner]
      properties:
        id:
          type: string
          format: uuid
          x-go-name: ID
          x-go-type: string
        name:
          type: string
        owner:
          $ref: '#/components/schemas/UserListItem'
        members:
          type: array
          items:
            $ref: '#/components/schemas/UserListItem'
        invite_code:
          type: string
        invite_code_expires_at:
          type: string
          description: 'JST。形式: YYYY-MM-DD HH:mm'
        is_owner:
          type: boolean
security:
  - bearerAuth: []